driver: Fix backward compatibility of lapic_timer_period symbol
It was not enough to only rename lapic_timer_period back to
lapic_timer_frequency for older kernels. First, our macro
RESOLVE_EXTERNAL_SYMBOL was not prepared to be provided a macro as well
and was still using the untranslated 5.3-name. And if we do not want to
break up that macro, we need to map lapic_timer_period_sym as well.

Reported-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Fixes: e7674ea7425f ("driver: Account for renaming of lapic_timer_frequency in 5.3")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>