hypervisor: introduce a common cache coloring interface
Add fundamental primitives to implement support for cache coloring.
The latter is performed by exploiting the mapping of memories flagged as
JAILHOUSE_MEM_COLORED and defined in the inmates' configurations (no
root-cell or hypervisor itself supported yet).
The interface defines the basic functions that have to be implemented in
order to support coloring for a specific architecture.

Define three main functions:
- coloring_paging_init: calculate which bit range in the address is useful
  for cache coloring, assuming it limited by the page size and the LLC
  way size as happens on arm;
- next_colored: calculate the next physical page address that conforms
  to a given colors selection and to the bits calculated in the previous
  function;
- coloring_cell_init: initialize cell colors assignement depending on its
  configuration;

LLC way size is required for coloring to work and it can be explicitly
defined in the root cell configuration. Add a helper function that
probes the way size from hardware if the llc_way_size is not set in the
root cell configuration.

Finally, define a function used to calculate the physical end address of
a given memory that is flagged as colored.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Signed-off-by: Marco Solieri <ms@xt3.it>