tools: ivshmem-demo: Map memory regions with correct sizes
ivshmem protocol does not describe a fixed size for the
rw, input and output regions. For each platform, the uio
driver will populate this information in the sysfs.

Extract the size from sysfs maps entries and use it for
mapping different regions.
This will make the demo generic such that it will work on
all platforms with different sizes for ivshmem.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
[Jan: terminate on error in of uio_read_mem_size, fix return type,
      process sscanf errors, close fd]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>