configs: imx8m: remove the physical address range contains GICD and GICRs
When enabling Jailhouse on the iMX8M platforms, the stage 2 translation
about IPA->PA includes the address range contains GICD and GICRs, which
should not be allowed because GICD or GICRs should not be accessed in
the root cell. In order to solve this problem, remove the physical
address range contains GICD and GICRs.

Signed-off-by: Alice Guo <alice.guo@nxp.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>