x86/cat.c: Fix off-by-one error
The variable cbm_max captures the position of the most significant bit
of the CBM, rather than the number of bits in the CBM. This should be
considered when checking non-root cell configurations to be valid,
otherwise the non-root cells can never access the top-most bit in the
CBM.

This off-by-one error ensured that the ROOT_COS CBM always had at least
one private bit in the CBM, preventing it from becoming empty. However,
empty ROOT_COS CBMs are well accounted for in the code.

Fixes: 3f04eb1753bb ("x86: Introduce Cache Allocation Technology support
for Intel CPUs")

Signed-off-by: Bram Hooimeijer <bram.hooimeijer@prodrive-technologies.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>