x86: Decode PCI config access violations
This has been a FAQ on the list way too long: Parsing the address port
register content, translating it into the target PCI device and config
space register is not for beginners and even not convenient for experts.
So add a dedicated printk for this case which allows to remove the
address port reporting from the remaining ones.

This now gives

FATAL: Invalid PCI config write, device 05:00.0, reg: 0x44, size: 2
...

analogously to the MMCONFIG access violations.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>