inmates: gic-demo: map led_reg
When I introduced MMU support for inmates, I forgot to map the memory
for LEDs. Fix that by simply mapping it.

While at it, move the initalisation of led_reg to the top of the routine
where it should actually be - it should be initialised before the timer
is armed.

The mapper will take care that the address will be huge page aligned,
and the size will be aligned to HUGE_PAGE_SIZE, so there's no need to
align anything on caller site.

Fixes: bca43a784716 ("inmates: arm-common: add MMU support")
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>