ci: Update kernel image to 5.10
The 4.13-based kernel image used for CI is not seriously aging,
including warnings with gcc-9 that is now coming via the Ubuntu 20.04 CI
image. Bump to 5.10 which should last for a while.

The kernel image is now fully built. That permits to check all symbols
the driver may need. Therefore, include all currently needed symbol
export patches in the generated kernel.

While updating the kernel configs, switch to defconfigs to reduce the
stored data.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>