README: arm64: disable ITS in qemu cmdline
GICv3s in recent QEMU versions have ITS (Interrupt Translation Service)
enabled by default, which Linux will make use of. We will crash when
trying to re-assigning a CPU back to the root cell, as the root
cell's Linux will try to access ITS regions when powering up a CPU.

As we don't support ITS in Jailhouse, disable ITS for the moment.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Reported-by: Stefan Huber <stefan.huber@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>