vmx: Add support for waitpkg instructions
Modern Intel CPUs provide a set of new instruction (tpause, umonitor,
umwait) which need to be permitted by the hypervisor. Otherwise, they
cause #UD. The Linux kernel uses tpause for implementing udelay since
kernel 5.8. So we will see crashes when denying this under Jailhouse.

The pattern to enable it is analogous to rdtscp, xsaves & Co.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>