x86: vtd: Drop source-id validation while parsing root cell IRTE
This was not complete, rather assumed SVT to be 01b and completely
ignored SQ. The SID test failed when the root cell chose SVT to be 10b
to describe a source behind a PCIe-PCI bridge, thus switching SID to
startbus/endbus encoding (now needed for QEMU).

Let's make it simple and assume that the root cell programmed a valid
source. Skipping the test will have no negative impact on isolation
between cells.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>