configs: x86: qemu: Move ivshmem devices behind hotplug-capable bridge
Recent QEMU disabled hotplugging support for the root bus. Linux will
still find the ivshmem devices we inject there so far, but it will fail
to allocate IRQs for it because the related irq domain was frozen in
size, given that hotplug is not possible.

Work around this issue by moving the virtual devices behind a PCIe-PCI
bridge that is hotplug-capable. Adding that bridge shuffles some memory
regions around, so adjust the configs accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>