tools: jailhouse-config-check: Add CPU set checks
When at least one non-root cell is provided, check if the root cell is
left with at least one core. If multiple non-root cells are listed, test
the those do not overlap in their CPU set. This ensures that the given
cell set can be instantiated at the same time.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>