configs: arm: emtrion-rzg1{h,m}: Fix config checker findings
In the emtrion-rzg1h, the APMU regions were overlapping. In the absence
if detailed information, keep the complete range mapped, analogously to
the emtrion-rzg1m.

The emtrion-rzg1h had a VSPDU region as duplicate of the VSPD2. Drop the
former.

While at it, remove JAILHOUSE_MEM_IO_32 where unneeded. This is only
relevant for sub-page regions.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>